Added a "Test Connection" button to each build step
